chore: split python cluster tests
